Fixed bug in GenericfMRIVolumeProcessingPipelineBatch.sh.
The script would print the desired resolution as "2.00", and that would confuse the script, which expected "2".  In that case, OneStepResampling.sh will run a flirt interpolation that ended up returning an image with different dimensions than the Atlas.
Now GenericfMRIVolumeProcessingPipelineBatch.sh removes trailing zeros in the resolution, so the behavior is the expected one.